# Avoid unnecessary __getitem__ in block() when chunks have correct dimensionality