@mrocklin 2020-02-12
This looks great to me.  Thanks @astrofrog ! Merging.

Also, I notice that this is your first code contribution to this repository.  Welcome!  I'm also personally excited to see you using Dask enough to find bugs.  I hope that the project is proving useful to you all in astronomy-land.